Extract NavBar as containers to be reused.
