+++ frontend/src/pages/CDA/CDA.js
@@ -13,8 +13,8 @@ import {
   initialRange,
   initialTrigger,
 } from "../../core/config";
-import NavBar from "../../common/components/NavBar";
 import Menu from "../../common/containers/Menu/Menu";
+import NavBar from "../../common/containers/NavBar/NavBar";
 import Entry from "./containers/Entry/Entry";
 import CountCardCont from "./containers/CountCard/CountCardCont";
 import Gallery from "./containers/Gallery/Gallery";
@@ -111,26 +111,26 @@ function CDA() {
     random: {
       name: "Random",
       icon: <FaRandom />,
-      style: { font: "text-2xl", width:"w-full" },
+      style: { font: "text-2xl", width: "w-full" },
       onClick: get_random_files,
     },
     save: {
       name: "Save",
       icon: <MdSave />,
-      style: { font: "text-2xl", width:"w-full" },
+      style: { font: "text-2xl", width: "w-full" },
       onClick: update_range,
     },
     reset: {
       name: "Reset",
       icon: <GrPowerReset />,
-      style: { font: "text-2xl", width:"w-full" },
+      style: { font: "text-2xl", width: "w-full" },
       onClick: get_trackbar,
     },
     process: {
       name: "Process",
       icon: <VscRunAll />,
       onClick: process_img,
-      style: { font: "text-2xl", width:"w-full" },
+      style: { font: "text-2xl", width: "w-full" },
       disabled: state === "started" || Number(entry.random) < 100,
     },
   };
